# Make sparse matrix constructor faster for Int32