Constructing a sparse matrix `sparse(I, J, V)` with `I` and `J` of type `Vector{Int32}` is slow due to conversions from Int64 -> Int32.

Example with matrix of order 10_000 with 30_000 nonzeros:

- Int32: 1.605ms (before) and 1.240ms (after)
- Int64: 1.303ms (before) and 1.306ms (after)