@haampie 2018-03-21
There's probably a few other functions with the same issue, I'll check!