@ViralBShah 2018-03-22
Do update this PR if you find more of these and we can merge it all together once you say it is ready.