@haampie 2018-03-22
Maybe it's best to merge this one already, cause going through SparseArrays in search of integer stuff leads me to other places as well.