@andreasnoack 2018-03-22
I'm wondering if using smaller integers during the calculations is just faster or if this is a type instability issue. For `Ti=Int32` `code_warntype` shows that `origcsrrowptri::Union{Int32, Int64}` so maybe that the root cause. Writing `Ti(1)` is a bit more verbose than `1` so if it is just the type instability it might be okay to do the other calculations with `1`.