@haampie 2018-03-22
Oh, I missed that! Yeah, I'd like to have just 1's everywhere, so I'll give it another take