@ViralBShah 2018-03-25
@ranjanan It would be good to see if this helps us improve circuitscape. We noticed that 32-bit ints in sparse matrices were not giving a meaningful speedup on 0.6.